# services/assistant-api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Assistant API",
    "description" : "Conversation and messaging service for jorisjonkers.dev",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://assistant.jorisjonkers.dev",
    "description" : "Production"
  }, {
    "url" : "http://localhost:8082",
    "description" : "Local development"
  } ],
  "security" : [ {
    "xUserId" : [ ]
  } ],
  "paths" : {
...
    "/api/v1/chat-sessions" : {
      "get" : {
        "tags" : [ "chat-session-controller" ],
        "operationId" : "list_4",
        "parameters" : [ {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ChatSessionResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "chat-session-controller" ],
        "operationId" : "create_4",
        "parameters" : [ {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StartChatSessionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChatSessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat-sessions/{id}/messages" : {
      "post" : {
        "tags" : [ "chat-session-controller" ],
        "operationId" : "appendMessage",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AppendChatMessageRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/conversations/{id}" : {
      "get" : {
        "tags" : [ "conversation-controller" ],
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConversationResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "conversation-controller" ],
        "operationId" : "archive",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/v1/chat-sessions/{id}" : {
      "get" : {
        "tags" : [ "chat-session-controller" ],
        "operationId" : "get_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "chat-session-controller" ],
        "operationId" : "archive_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/workspaces/{workspaceId}/sessions/{sessionId}" : {
      "delete" : {
        "tags" : [ "agent-session-controller" ],
        "operationId" : "stop",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "sessionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ConversationResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "createdAt", "id", "status", "title", "updatedAt", "userId" ]
      },
...
      "StartChatSessionRequest" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : [ "string", "null" ],
            "maxLength" : 120,
            "minLength" : 0
          },
          "kind" : {
            "type" : [ "string", "null" ],
            "enum" : [ "PLAIN", "KNOWLEDGE" ]
          }
        }
      },
      "ChatSessionResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "createdAt", "id", "kind", "status", "updatedAt", "userId" ]
      },
      "AppendChatMessageRequest" : {
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "string",
            "minLength" : 1
          },
          "role" : {
            "type" : "string",
            "enum" : [ "USER", "ASSISTANT", "SYSTEM" ]
          }
        },
        "required" : [ "body", "role" ]
      },
      "ChatMessageResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "sessionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "role" : {
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "body", "createdAt", "id", "role", "sessionId" ]
      },
...
    },
    "securitySchemes" : {
      "xUserId" : {
        "type" : "apiKey",
        "description" : "User identifier injected by Traefik forward-auth",
        "name" : "X-User-Id",
        "in" : "header"
      }
    }
  }
}